remove old files + move following
